Fixup non-conflicting files

- GAME_START packet contents
- Server: Bring back endGame() call
- Use Ticked packets for SHOOT and USE_ITEM. Server requires all unreliable packets to be Ticked.
- Silence warnings about unused update(dt) parameters

+++ src/HealthBar.h
@@ -18,7 +18,7 @@ class HealthBar : public sf::Drawable, public sf::Transformable
 	int getHealth() const;
 	int getMaxHealth() const;
 
-	void update(float dt);
+	void update(int32_t dt);
 	void healthCallback(int health, int max);
 
   private:
